Extract default hook return values
